pyproject.toml
[project]
name = "publications-rdi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "marimo>=0.17.0",
    "polars>=1.34.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
]
